Ui + attaques (Scriptable)

+++ Rumble In Chains/Assets/Scripts/Attacks/HitboxCapsule.cs
@@ -14,7 +14,7 @@ public class HitboxCapsule : Hitbox // !!!
     public Vector2 CenterOfTheSecondSphere { get => _centerOfTheSecondSphere; }
     public float Radius { get => _radius; }
 
-    public HitboxCapsule(float damage, float stunFactor, float startUpTiming, float durationOfHitbox, Vector2 expulsion, Vector2 centerOfTheFirstSphere, Vector2 centerOfTheSecondSphere, float radius) : base(damage, stunFactor, startUpTiming, durationOfHitbox, expulsion)
+    public HitboxCapsule(float damage, float stunFactor, int startUpTiming, int durationOfHitbox, Vector2 expulsion, Vector2 centerOfTheFirstSphere, Vector2 centerOfTheSecondSphere, float radius) : base(damage, stunFactor, startUpTiming, durationOfHitbox, expulsion)
     {
         _centerOfTheFirstSphere = centerOfTheFirstSphere;
         _centerOfTheSecondSphere = centerOfTheSecondSphere;
